backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
asyncpg
python-dotenv
